Separate entrypoint and application
